Extract protocol to handle quantity logic.
